fileserver/requirements.txt
click==6.7
Flask==0.12.2
Flask-PyMongo==0.5.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pymongo==3.4.0
python-magic==0.4.13
Werkzeug==0.12.2
